Split abstract states to separate module
